dockercmp_flask_blueprint/requirements.txt
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
cssmin==0.2.0
cycler==0.11.0
Flask==2.2.2
Flask-Assets==2.0
fonttools==4.37.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
webassets==2.0
Werkzeug==2.2.2
